+++ src/main/java/net/arcanamod/client/render/tiles/AspectBookshelfTileEntityRenderer.java
@@ -1,0 +1,5 @@
+package net.arcanamod.client.render.tiles;
+
+public class AspectBookshelfTileEntityRenderer {
+    // TODO: All
+}
